fix(PageHeader): correct dependencies, move stories
