+++ stories/ui-page-header/.storybook/main.js
@@ -1,0 +1,9 @@
+const addonConfigs = require('../../../.storybook/utilities/addon-configs');
+
+module.exports = {
+  addons: [
+    ...addonConfigs.addonsNoEssentials,
+    ...addonConfigs.essentialsNoDocs,
+  ],
+  stories: ['../**/*.stories.@(js|ts|mdx)'],
+};
